sync/requirements.txt
sqlalchemy>=2.0.0
syncthing-python-client>=0.8.0
tqdm>=4.65.0
pydantic>=2.0.0
python-dotenv>=1.0.0 